fix: Corrijo bug para que el misil explote correctamente al lanzarse cerca de una pared + Corrijo direccion e intensidad de la explosion

+++ src/client/client_explosion.cpp
@@ -26,7 +26,7 @@ void Explosion::render(int frame) {
       //std::cout << "fraction n° " << i << ": " << ray_fractions[i] << std::endl;
       float radius_fraction = ray_fractions[i] * radius;
       SDL2pp::Point center (pos_x, pos_y);
-      SDL2pp::Point ray_dir ( radius_fraction * sin(angle_rad), -radius_fraction * cos(angle_rad) );
+      SDL2pp::Point ray_dir ( radius_fraction * sinf(angle_rad), -radius_fraction * cosf(angle_rad) );
       SDL2pp::Point ray_end = (center + ray_dir);
       renderer.DrawLine(center, ray_end);
     }
